requirements.txt
openai
python-dotenv
boto3
pygame

sounddevice
soundfile

scipy
wavio

# sudo apt-get install ffmpeg
pydub